packages/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
asyncpg>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
redis[hiredis]>=5.2.0
fastembed>=0.4.0
mcp>=1.0.0
firebase-admin>=6.6.0
pyjwt>=2.9.0
bcrypt>=4.2.0
httpx>=0.28.0
google-cloud-tasks>=2.16.0
google-cloud-secret-manager>=2.21.0

# Dev / test
pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
mypy>=1.13.0
asyncpg-stubs>=0.31.0
